@@ -14,9 +14,9 @@
   <name>lists-service</name>
   <description>REST and GraphQL APIs for managing and searching species lists</description>
   <properties>
-    <java.version>17</java.version>
+    <java.version>21</java.version>
     <spotless-maven-plugin.version>2.22.8</spotless-maven-plugin.version>
-    <spring-boot-maven-plugin.version>3.0.2</spring-boot-maven-plugin.version>
+    <spring-boot-maven-plugin.version>3.1.6</spring-boot-maven-plugin.version>
   </properties>
 
   <repositories>
@@ -56,12 +56,12 @@
     <dependency>
       <groupId>au.org.ala</groupId>
       <artifactId>ala-ws-spring-security</artifactId>
-      <version>7.2.0-SNAPSHOT</version>
+      <version>7.3.0-SNAPSHOT</version>
     </dependency>
     <dependency>
       <groupId>au.org.ala</groupId>
       <artifactId>ala-ws-security</artifactId>
-      <version>7.2.0-SNAPSHOT</version>
+      <version>7.3.0-SNAPSHOT</version>
     </dependency>
     <dependency>
       <groupId>com.fasterxml.jackson.module</groupId>
@@ -107,10 +107,6 @@
       <artifactId>lombok</artifactId>
     </dependency>
     <dependency>
-      <groupId>org.springframework.graphql</groupId>
-      <artifactId>spring-graphql-test</artifactId>
-    </dependency>
-    <dependency>
       <groupId>org.gbif</groupId>
       <artifactId>dwca-io</artifactId>
       <version>2.16</version>
@@ -138,6 +134,22 @@
   </dependencies>
   <build>
     <plugins>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-compiler-plugin</artifactId>
+        <version>3.11.0</version>
+        <configuration>
+          <source>21</source>
+          <target>21</target>
+          <annotationProcessorPaths>
+            <path>
+              <groupId>org.projectlombok</groupId>
+              <artifactId>lombok</artifactId>
+              <version>1.18.30</version>
+            </path>
+          </annotationProcessorPaths>
+        </configuration>
+      </plugin>
       <plugin>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-maven-plugin</artifactId>
