@@ -7,7 +7,8 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.annotation.Order;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.Customizer;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.web.SecurityFilterChain;
@@ -17,8 +18,8 @@ import org.springframework.security.web.firewall.HttpFirewall;
 
 @Configuration
 @EnableWebSecurity
-@ComponentScan(basePackages = {"au.org.ala.ws.security"})
-@EnableGlobalMethodSecurity(securedEnabled = true)
+@ComponentScan(basePackages = {"au.org.ala.ws.security", "au.org.ala.security.common"})
+@EnableMethodSecurity(securedEnabled = true)
 @EnableCaching
 @Order(1)
 public class SecurityConfig {
@@ -29,9 +30,34 @@ public class SecurityConfig {
   public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
 
     http.addFilterBefore(alaWebServiceAuthFilter, BasicAuthenticationFilter.class);
-    http.authorizeRequests()
+    http.authorizeHttpRequests(auth -> auth
             .requestMatchers("/", "/graphql", "/ingest", "/graphiql", "/v1/species/**", "/**")
-            .permitAll();
+            .permitAll());
+
+    /*
+      Configure security headers to address vulnerabilities:
+      - Fix Missing HTTP Strict Transport Security Policy (Medium severity)
+      - Fix Missing 'X-Frame-Options' Header (Low severity) - prevents clickjacking
+      - Fix Missing 'X-Content-Type-Options' Header (Low severity) - prevents MIME sniffing
+      - Fix Missing Content Security Policy (Low severity) - prevents XSS attacks
+     */
+    http.headers(headers -> headers
+        .httpStrictTransportSecurity(hstsConfig -> hstsConfig
+            .maxAgeInSeconds(31536000) // 1 year
+            .includeSubDomains(true)
+            .preload(true))
+        .frameOptions(frameOptions -> frameOptions.deny())
+        .contentTypeOptions(Customizer.withDefaults())
+        .contentSecurityPolicy(csp -> csp.policyDirectives(
+            "default-src 'self'; " +
+                "script-src 'self' 'unsafe-inline' 'unsafe-eval'; " +
+                "style-src 'self' 'unsafe-inline'; " +
+                "img-src 'self' data: https:; " +
+                "font-src 'self' data:; " +
+                "connect-src 'self'; " +
+                "frame-ancestors 'none'; " +
+                "base-uri 'self'")));
+
     return http.csrf(httpSecurityCsrfConfigurer -> httpSecurityCsrfConfigurer.disable()).build();
   }
 
