@@ -1,5 +1,5 @@
 /* eslint-disable react-hooks/exhaustive-deps */
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { Facet, KV, queries, SpeciesListPage, useGQLQuery } from '#/api';
 import {
   ActionIcon,
   Box,
@@ -15,6 +15,7 @@ import {
   Paper,
   SegmentedControl,
   Select,
+  Skeleton,
   Space,
   Stack,
   Table,
@@ -23,35 +24,48 @@ import {
   Title,
   Tooltip,
 } from '@mantine/core';
-import { useMediaQuery } from '@mantine/hooks';
-import { useGQLQuery, queries, SpeciesListPage, KV, Facet } from '#/api';
-import { useDebouncedValue, useDocumentTitle } from '@mantine/hooks';
-import { FormattedMessage, FormattedNumber, useIntl } from 'react-intl';
+import {
+  useDebouncedValue,
+  useDocumentTitle,
+  useMediaQuery,
+} from '@mantine/hooks';
 import {
   parseAsBoolean,
   parseAsInteger,
   parseAsString,
   useQueryState,
 } from 'nuqs';
+import { useCallback, useEffect, useMemo, useState } from 'react';
+import { FormattedMessage, FormattedNumber, useIntl } from 'react-intl';
 
 // Icons
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faEye, faEyeSlash, faXmark, faMagnifyingGlass, faCode } from '@fortawesome/free-solid-svg-icons';
 import { StopIcon } from '@atlasoflivingaustralia/ala-mantine';
+import {
+  faCode,
+  faEye,
+  faEyeSlash,
+  faMagnifyingGlass,
+  faXmark,
+} from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 // Project components
 import { Message } from '#/components/Message';
 import { ListRow } from './components/ListRow';
 
 // Helpers
+import {
+  ActiveFilters,
+  FiltersSection,
+  ToggleFiltersButton,
+} from '#/components/FiltersSection';
 import { getErrorMessage, parseAsFilters } from '#/helpers';
 import { useALA } from '#/helpers/context/useALA';
-import { ToggleFiltersButton, ActiveFilters, FiltersSection } from '#/components/FiltersSection';
 
 // Styles
-import classes from './classes/index.module.css';
-import { Breadcrumbs } from '../Dashboard/components/Breadcrumbs';
 import { Link } from 'react-router';
+import { Breadcrumbs } from '../Dashboard/components/Breadcrumbs';
+import classes from './classes/index.module.css';
 
 interface HomeQuery {
   lists: SpeciesListPage;
@@ -59,6 +73,7 @@ interface HomeQuery {
 }
 
 const sortField = [
+  'relevance_desc',
   'lastUpdated_desc',
   'lastUpdated_asc',
   'title_asc',
@@ -90,7 +105,7 @@ function Home() {
   );
   const [sort, setSort] = useQueryState<string>(
     'sort',
-    parseAsString.withDefault('lastUpdated')
+    parseAsString.withDefault('relevance')
   );
   const [dir, setDir] = useQueryState<string>(
     'dir',
@@ -105,7 +120,7 @@ function Home() {
     parseAsBoolean.withDefault(false)
   );
   const [filters, setFilters] = useQueryState<KV[]>(
-    'filters', 
+    'filters',
     parseAsFilters // Note: adding `.withDefault([])` causes infinite loop (bug in nuqs v2.4.1 ??)
   );
 
@@ -113,11 +128,14 @@ function Home() {
   const [refresh, setRefresh] = useState<boolean>(false);
 
   // Filters display state
-  const [hidefilters, setHideFilters] = useQueryState<boolean>('hideFilters', parseAsBoolean.withDefault(false)); 
+  const [hidefilters, setHideFilters] = useQueryState<boolean>(
+    'hideFilters',
+    parseAsBoolean.withDefault(false)
+  );
   const toggleFilters = () => setHideFilters((o) => !o);
 
   const ala = useALA();
-  const { data, error, update } = useGQLQuery<HomeQuery>(
+  const { data, error, loading, update } = useGQLQuery<HomeQuery>(
     queries.QUERY_LISTS_SEARCH,
     {
       searchQuery: search,
@@ -159,25 +177,28 @@ function Home() {
     // Hide filters for mobile devices
     setHideFilters(isMobile);
   }, [isMobile]);
-  
 
   // Retry handler
   const handleRetry = useCallback(() => {
     setPage(0);
     setSize(10);
-    setSort('title');
+    setSort('relevance_desc');
     setDir('desc');
-    setSearch('lastUpdated');
+    setSearch('');
     setView('public');
     setIsUser(false);
     setRefresh(!refresh);
   }, [refresh]);
 
-  const sortOptions = useMemo(() => 
-    sortField.map((key) => ({
-      label: intl.formatMessage({ id: key, defaultMessage: key.replace(/_/g, ' ') }),
-      value: key,
-    })), 
+  const sortOptions = useMemo(
+    () =>
+      sortField.map((key) => ({
+        label: intl.formatMessage({
+          id: key,
+          defaultMessage: key.replace(/_/g, ' '),
+        }),
+        value: key,
+      })),
     [intl]
   );
 
@@ -203,12 +224,10 @@ function Home() {
     [filters]
   );
 
-  const resetFilters = useCallback(
-    () => {
-      setPage(0); // Reset 'page' when filters are reset
-      setFilters([]);
-    },[filters]
-  );
+  const resetFilters = useCallback(() => {
+    setPage(0); // Reset 'page' when filters are reset
+    setFilters([]);
+  }, [filters]);
 
   const labels = useMemo(
     () => [
@@ -217,7 +236,9 @@ function Home() {
         label: (
           <Center style={{ gap: 10 }}>
             <FontAwesomeIcon icon={faEye} fontSize={14} />
-            <span><FormattedMessage id='public.label' defaultMessage='Public' /></span>
+            <span>
+              <FormattedMessage id='public.label' defaultMessage='Public' />
+            </span>
           </Center>
         ),
       },
@@ -226,7 +247,9 @@ function Home() {
         label: (
           <Center style={{ gap: 10 }}>
             <FontAwesomeIcon icon={faEyeSlash} fontSize={14} />
-            <span><FormattedMessage id='private.label' defaultMessage='Private' /></span>
+            <span>
+              <FormattedMessage id='private.label' defaultMessage='Private' />
+            </span>
           </Center>
         ),
       },
@@ -241,20 +264,36 @@ function Home() {
       <Container fluid className={classes.speciesHeader}>
         <Grid>
           <Grid.Col span={12}>
-            <Breadcrumbs listTitle={undefined}/>
+            <Breadcrumbs listTitle={undefined} />
           </Grid.Col>
           <Grid.Col span={12}>
-            <Title order={3} classNames={{ root: classes.title }} >
-              <FormattedMessage id='lists.title' defaultMessage='Species Lists' />
+            <Title order={3} classNames={{ root: classes.title }}>
+              <FormattedMessage
+                id='lists.title'
+                defaultMessage='Species Lists'
+              />
             </Title>
           </Grid.Col>
           <Grid.Col span={9}>
-            <Title order={5} classNames={{ root: classes.subtitle }} >
-              <FormattedMessage id='lists.subtitle' defaultMessage='A tool for finding species checklists' />
+            <Title order={5} classNames={{ root: classes.subtitle }}>
+              <FormattedMessage
+                id='lists.subtitle'
+                defaultMessage='A tool for finding species checklists'
+              />
             </Title>
           </Grid.Col>
-          <Grid.Col span={3} style={{ display: 'flex', justifyContent: 'flex-end' }}>
-            <Tooltip label={intl.formatMessage({ id: 'openapi.button.title' })} position='left' withArrow multiline w={200} opacity={0.8}>
+          <Grid.Col
+            span={3}
+            style={{ display: 'flex', justifyContent: 'flex-end' }}
+          >
+            <Tooltip
+              label={intl.formatMessage({ id: 'openapi.button.title' })}
+              position='left'
+              withArrow
+              multiline
+              w={200}
+              opacity={0.8}
+            >
               <Button
                 variant='default'
                 radius='xl'
@@ -266,7 +305,10 @@ function Home() {
                 aria-label={intl.formatMessage({ id: 'openapi.button.title' })}
                 leftSection={<FontAwesomeIcon icon={faCode} />}
               >
-                <FormattedMessage id='openapi.button.label' defaultMessage='OpenAPI'/>
+                <FormattedMessage
+                  id='openapi.button.label'
+                  defaultMessage='OpenAPI'
+                />
               </Button>
             </Tooltip>
           </Grid.Col>
@@ -276,8 +318,11 @@ function Home() {
         <Grid>
           <Grid.Col span={12}>
             <Group>
-              { !isMobile && (
-                <ToggleFiltersButton toggleFilters={toggleFilters} hidefilters={hidefilters} />
+              {!isMobile && (
+                <ToggleFiltersButton
+                  toggleFilters={toggleFilters}
+                  hidefilters={hidefilters}
+                />
               )}
               <TextInput
                 style={{ flexGrow: 1 }}
@@ -287,25 +332,40 @@ function Home() {
                   setPage(0); // Reset 'page' when search is changed
                   setSearch(event.currentTarget.value);
                 }}
-                placeholder={intl.formatMessage({ id: 'search.input.placeholder', defaultMessage: 'Search lists by name or taxa' })}
+                placeholder={intl.formatMessage({
+                  id: 'search.input.placeholder',
+                  defaultMessage: 'Search lists by name or taxa',
+                })}
                 w={200}
-                leftSection={<FontAwesomeIcon icon={faMagnifyingGlass} fontSize={16} stroke='2' />}
+                leftSection={
+                  <FontAwesomeIcon
+                    icon={faMagnifyingGlass}
+                    fontSize={16}
+                    stroke='2'
+                  />
+                }
                 rightSection={
                   <ActionIcon
                     radius='sm'
-                    variant="transparent"
+                    variant='transparent'
                     size='xs'
-                    title={intl.formatMessage({ id: 'search.clear.label', defaultMessage: 'Clear search' })}
-                    aria-label={intl.formatMessage({ id: 'search.clear.label', defaultMessage: 'Clear search' })}
+                    title={intl.formatMessage({
+                      id: 'search.clear.label',
+                      defaultMessage: 'Clear search',
+                    })}
+                    aria-label={intl.formatMessage({
+                      id: 'search.clear.label',
+                      defaultMessage: 'Clear search',
+                    })}
                     disabled={search.length === 0}
                     onClick={() => setSearch('')}
                     style={{ marginLeft: 5, marginRight: 10 }}
                   >
-                  <FontAwesomeIcon icon={faXmark} fontSize={20} />
+                    <FontAwesomeIcon icon={faXmark} fontSize={20} />
                   </ActionIcon>
                 }
               />
-              {ala.isAuthenticated &&  (
+              {ala.isAuthenticated && (
                 <>
                   <SegmentedControl
                     disabled={!data || hasError}
@@ -315,8 +375,13 @@ function Home() {
                     data={labels}
                   />
                   <Checkbox
-                    label={<FormattedMessage id='myLists.label' defaultMessage='My Lists' />}
-                    checked={isUser}  
+                    label={
+                      <FormattedMessage
+                        id='myLists.label'
+                        defaultMessage='My Lists'
+                      />
+                    }
+                    checked={isUser}
                     size='sm'
                     onChange={(e) => setIsUser(e.currentTarget.checked)}
                     classNames={{ label: classes.myListsLabel }}
@@ -326,7 +391,9 @@ function Home() {
               <Select
                 w={235}
                 value={`${sort}_${dir}`}
-                label={<FormattedMessage id='sort.label' defaultMessage='Sort by' />}
+                label={
+                  <FormattedMessage id='sort.label' defaultMessage='Sort by' />
+                }
                 withCheckIcon={true}
                 data={sortOptions}
                 classNames={{
@@ -334,13 +401,16 @@ function Home() {
                   label: classes.sortSelectLabel,
                 }}
                 onChange={(value: string | null) => {
-                  const [sort, dir] = (value || 'title_asc').split('_');
+                  const [sort, dir] = (value || 'relevance_desc').split('_');
                   setSort(sort);
                   setDir(dir);
                   setPage(0);
                 }}
                 disabled={!data || hasError}
-                aria-label={intl.formatMessage({ id: 'sort.field.ariaLabel', defaultMessage: 'Select field to sort results' })} 
+                aria-label={intl.formatMessage({
+                  id: 'sort.field.ariaLabel',
+                  defaultMessage: 'Select field to sort results',
+                })}
               />
               <Select
                 w={110}
@@ -351,71 +421,125 @@ function Home() {
                   value,
                 }))}
                 disabled={!data || hasError}
-                aria-label={intl.formatMessage({ id: 'page.size.ariaLabel', defaultMessage: 'Select number of results' })} 
+                aria-label={intl.formatMessage({
+                  id: 'page.size.ariaLabel',
+                  defaultMessage: 'Select number of results',
+                })}
               />
-              { isMobile && (
-                <ToggleFiltersButton toggleFilters={toggleFilters} hidefilters={hidefilters} />
+              {isMobile && (
+                <ToggleFiltersButton
+                  toggleFilters={toggleFilters}
+                  hidefilters={hidefilters}
+                />
               )}
             </Group>
           </Grid.Col>
           {!hidefilters && (
             <Grid.Col span={{ base: 12, sm: 4, md: 3, lg: 2 }} mt={16}>
               <Collapse in={!hidefilters}>
-                  {/* Filters appear here */}
-                  <FiltersSection
-                    facets={data?.facets || []}
-                    active={filters || []}
-                    onSelect={handleFilterClick}
-                    onReset={() => {setFilters([]); setPage(0);}}
-                  />
+                {/* Filters appear here */}
+                <FiltersSection
+                  facets={data?.facets || []}
+                  active={filters || []}
+                  onSelect={handleFilterClick}
+                  onReset={() => {
+                    setFilters([]);
+                    setPage(0);
+                  }}
+                />
               </Collapse>
             </Grid.Col>
           )}
-          <Grid.Col span={{ base: 12, sm: hidefilters ? 12 : 8, md: hidefilters ? 12 : 9, lg: hidefilters ? 12 : 10 }}>
+          <Grid.Col
+            span={{
+              base: 12,
+              sm: hidefilters ? 12 : 8,
+              md: hidefilters ? 12 : 9,
+              lg: hidefilters ? 12 : 10,
+            }}
+          >
             <Grid.Col span={12}>
               {error && (
                 <Message
-                  title={intl.formatMessage({ id: 'error.error.title', defaultMessage: 'An error occurred' })} 
+                  title={intl.formatMessage({
+                    id: 'error.error.title',
+                    defaultMessage: 'An error occurred',
+                  })}
                   subtitle={getErrorMessage(error)}
                   icon={<StopIcon size={18} />}
-                  action={intl.formatMessage({ id: 'error.action.label', defaultMessage: 'Retry' })} 
+                  action={intl.formatMessage({
+                    id: 'error.action.label',
+                    defaultMessage: 'Retry',
+                  })}
                   onAction={handleRetry}
                 />
               )}
               {!error && (
                 <>
                   <Box ml={10} mb={0} mt={5}>
-                    { (totalElements && totalElements > 0) ? (
+                    {totalElements && totalElements > 0 ? (
                       <>
-                        <Text size='sm' mb={6} mt={4} className={classes.resultsSummary} component='span'>
-                          <FormattedMessage id='results.showing' defaultMessage='Showing' /> {' '}
+                        <Text
+                          size='sm'
+                          mb={6}
+                          mt={4}
+                          className={classes.resultsSummary}
+                          component='span'
+                        >
+                          <FormattedMessage
+                            id='results.showing'
+                            defaultMessage='Showing'
+                          />{' '}
                           {(realPage - 1) * size + 1}-
-                          {Math.min((realPage - 1) * size + size, totalElements || 0)} of {' '}
-                          <FormattedNumber value={totalElements || 0} /> {' '}
-                          <FormattedMessage id='results.records' defaultMessage='records' />
-                          { filters && filters.length > 0 && (
-                            <><Space w={5} />–<Space  w={2} /></>
+                          {Math.min(
+                            (realPage - 1) * size + size,
+                            totalElements || 0
+                          )}{' '}
+                          of <FormattedNumber value={totalElements || 0} />{' '}
+                          <FormattedMessage
+                            id='results.records'
+                            defaultMessage='records'
+                          />
+                          {filters && filters.length > 0 && (
+                            <>
+                              <Space w={5} />–<Space w={2} />
+                            </>
                           )}
                         </Text>
                       </>
                     ) : (
-                      <Text size='sm' mb={6} mt={4} className={classes.resultsSummary} component='span'>
-                        <FormattedMessage id='results.noRecords' defaultMessage='No records found' />
-                        { search && search.length > 0 ? (
-                          <>{' '} <FormattedMessage id='results.for' defaultMessage='for' /> "{search}"{' '}</>
-                        ) : (
-                          <>{' '}</>
-                        )}
-                        { filters && filters.length > 0 && (
-                          <>with{' '}</>
-                        )}
-                      </Text>
-                    )}
-                    { filters && filters.length > 0 && (
-                      <Paper 
-                        ml={4} 
-                        className={classes.resultsSummary}
+                      <Skeleton
+                        mb={6}
+                        mt={4}
+                        visible={!totalElements && loading}
                       >
+                        <Text
+                          size='sm'
+                          className={classes.resultsSummary}
+                          component='span'
+                        >
+                          <FormattedMessage
+                            id='results.noRecords'
+                            defaultMessage='No records found'
+                          />
+                          {search && search.length > 0 ? (
+                            <>
+                              {' '}
+                              <FormattedMessage
+                                id='results.for'
+                                defaultMessage='for'
+                              />{' '}
+                              "{search}"{' '}
+                            </>
+                          ) : (
+                            <> </>
+                          )}
+                          {filters && filters.length > 0 && <>with </>}
+                        </Text>
+                      </Skeleton>
+                    )}
+                    {filters && filters.length > 0 && (
+                      <Paper ml={4} className={classes.resultsSummary}>
                         <ActiveFilters
                           active={filters}
                           handleFilterClick={handleFilterClick}
@@ -424,10 +548,16 @@ function Home() {
                       </Paper>
                     )}
                   </Box>
-                  <Table striped={false} withRowBorders className= {classes.resultsTable}>
+                  <Table
+                    striped={false}
+                    withRowBorders
+                    className={classes.resultsTable}
+                  >
                     <Table.Tbody>
                       {content
-                        ? content.map((list) => <ListRow key={list.id} list={list} />)
+                        ? content.map((list) => (
+                            <ListRow key={list.id} list={list} />
+                          ))
                         : Array.from(Array(size).keys()).map((key) => (
                             <ListRow key={key} />
                           ))}
@@ -435,13 +565,20 @@ function Home() {
                   </Table>
                 </>
               )}
-              {totalElements === 0 && 
-                <Message 
-                  title={intl.formatMessage({ id: 'error.noListsFound.title', defaultMessage: 'No lists found' })} 
-                  subtitle={intl.formatMessage({ id: 'error.noListsFound.subTitle', defaultMessage: 'Try removing a filter or searching with a different query' })} 
+              {totalElements === 0 && (
+                <Message
+                  title={intl.formatMessage({
+                    id: 'error.noListsFound.title',
+                    defaultMessage: 'No lists found',
+                  })}
+                  subtitle={intl.formatMessage({
+                    id: 'error.noListsFound.subTitle',
+                    defaultMessage:
+                      'Try removing a filter or searching with a different query',
+                  })}
                 />
-              }
-              <Stack align="center" justify="center" gap="xs" w="100%" py="xl">
+              )}
+              <Stack align='center' justify='center' gap='xs' w='100%' py='xl'>
                 <Pagination
                   disabled={(totalPages || 0) < 1 || hasError}
                   value={realPage}
